+++ python/ray/tests/BUILD
@@ -111,6 +111,7 @@ py_test_module_list(
     "test_memory_pressure.py",
     "test_metrics.py",
     "test_task_metrics.py",
+    "test_task_metrics_reconstruction.py",
     "test_actor_state_metrics.py",
     "test_placement_group_metrics.py",
     "test_object_store_metrics.py",
